MenuGuide/MenuButton/Play Menu sprite added
Completed MenuGuide, Menu Buttons.

Menu Guide is a class meant to organize and bind certain menu_selects via array. it has a toggle method.

Toggle to activate and deactivate them to to resemble the game switching menus between characters.

The clear method resets the contents of the menu guide array and deletes via WM. for future use in transition between phases..

Control toggle is a setter if need be.

Much of the button overall design accept for callback is from menu_select due to the similarity, callback is already apart of the event handler and so will be modify or possibly have a function as an argument for each button case.